requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
python-dotenv>=1.0.0
websockets>=11.0.0
httpx>=0.24.0
openai>=1.0.0
google-generativeai>=0.3.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
aiosqlite>=0.19.0
